requirements.txt
# YouTube Channel Predictor - Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
# or
# torch>=2.0.0

# YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
youtube-transcript-api>=0.6.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Jupyter (optional, for development)
jupyter>=1.0.0
ipykernel>=6.25.0

# Hugging Face (for data storage)
huggingface-hub>=0.17.0
